requirements.txt
aniso8601==3.0.2
autopep8==1.4.3
Click==7.0
Flask==1.0.2
Flask-GraphQL==2.0.0
Flask-SQLAlchemy==2.3.2
graphene==2.1.3
graphene-sqlalchemy==2.1.1
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
promise==2.2.1
pycodestyle==2.5.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.2
uWSGI==2.0.18
Werkzeug==0.15.2
